+++ CHANGELOG.md
@@ -11,6 +11,7 @@
 - feature: install cline
 - feature: install continue
 - feature: install cursor cli and gui
+- feature: install frame0
 
 ## [1.4.0] - 2026-07-24